Added a folder for all solver related functions
