+++ src/restruct_poc/solver/ConstraintInput.hpp
