+++ src/restruct_poc/solver/PredictNextState.hpp
