+++ tests/unit_tests/restruct_poc/test_beams.cpp
@@ -14,7 +14,7 @@
 #include "src/restruct_poc/CreateBeams.hpp"
 #include "src/restruct_poc/UpdateState.hpp"
 #include "src/restruct_poc/beams.hpp"
-#include "src/restruct_poc/state.hpp"
+#include "src/restruct_poc/solver/State.hpp"
 #include "tests/unit_tests/gen_alpha_poc/test_utilities.h"
 
 namespace openturbine::restruct_poc::tests {
