+++ tests/unit_tests/restruct_poc/test_cantilever_beam.cpp
@@ -9,10 +9,10 @@
 #include "src/restruct_poc/BeamSection.hpp"
 #include "src/restruct_poc/BeamsInput.hpp"
 #include "src/restruct_poc/CreateBeams.hpp"
-#include "src/restruct_poc/InitializeConstraints.hpp"
-#include "src/restruct_poc/Solver.hpp"
-#include "src/restruct_poc/Step.hpp"
 #include "src/restruct_poc/beams.hpp"
+#include "src/restruct_poc/solver/InitializeConstraints.hpp"
+#include "src/restruct_poc/solver/Solver.hpp"
+#include "src/restruct_poc/solver/Step.hpp"
 #include "tests/unit_tests/gen_alpha_poc/test_utilities.h"
 
 using BeamQuadrature = std::vector<std::array<double, 2>>;
